[clang][deps] Consolidate types into new `DependencyConsumer.h`
